Change border_edges from unordered_set to set
An attempt to unify behavior across compiler and OS's. With unordered_set,
the order in which we operate on border edges depends both on memory
allocation behavior *and* hashing functionality. With set, we're only
dependent on memory allocation behavior, removing one observed variable
(in mac, it became apparent that the hashing produced a different order
than on ubuntu.)